Removes tracking composer.lock.
Why?

We run and test the repo with different PHP versions which impacts the composer package and dependencies versions. Removing composer.lock simplifies this process.

How?

The lock file is locked to a specific version. For example, if I'm developing or testing with PHP 7.4, my lock file is different from if it was committed with PHP 5.6.